+++ tests/testthat/test-kable-coloring.R
@@ -1,5 +1,4 @@
 
-context("kable_coloring")
 
 testthat::test_that("kable_coloring basics", {
    # simple test vectors